Fix stop all processes when a process is killed
When a process is killed an exception is raised (killing a process
should not be needed, ideally all Clara processes should terminate
nicely). The exception was not caught and some of the process were never
requested to stop.

Fix this by catching the exception inside the loop, so all the processes
are stopped.