Ensure list of packages is updated
For some reason, Ubuntu repositories like to remove old version of
packages. If the list of packages is not updated, it will likely
generate an error when trying to provision the machine.

Fix with a shell script provisioner that will run before Puppet. This
script will execute `apt-get update` when the machine is created. Then,
it will run the command again only if the list has not been updated in
one week.