Fix process exit
When the Clara platform or DPE are launched from a shell script, and the
process is terminated by clara-manager, only the parent shell process
will finish, and the Clara process will keep running.

Install and use Google's psutil module to terminate both the parent
shell process and its children. This will finish the Clara child
process.

Don't terminate the parent process when there is a child. Given that
with Clara the parent is a shell script, it will finish after the Clara
process terminated. Trying to terminate the parent results in the
process being killed (which returns an error to the client).