Fix colors of clara_install script
Wrap print to reset the color, because it is not reset automatically
when the script is run through ssh.

The autoreset keyword cannot be removed from init(), or the output will
be totally crazy: the output of all the commands will be shown first and
then all the colored messages of the script.