plot: do not require YAML file to parse services
Update the parser to ensure the benchmark block is parsed in order, and
use all services as reported by the orchestrator.

The YAML was used to filter parsed lines that contained valid service
names, but with the new regexes there is no risk to parse a line that
doesn't contain a service benchmark result.